requirements.txt
Pillow>=8.4.0,<8.5.0
